+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Escape executed commands in logs, making them paste-able
 - Oversubscription now takes minimum of old and of new calculation
 - Autoplace API answer now contains object refs to nodes and storagepools
+- Make-Available will now retry auto-place with no restrictions on diskless resources
 
 ### Fixed